# openapi.yaml
...
openapi: 3.1.0
info:
  title: Mapvest API
  version: 0.1.0-alpha.0
  description: Mapvest turns places and objects into investable tickers. All request and response shapes are generated from zod schemas in `packages/core/src/schemas` — never hand-edit `openapi.yaml`.
  contact:
    name: Mapvest
    url: https://github.com/jawauntb/mapvest
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
  - url: https://mapvest-api.up.railway.app
    description: Railway (staging/production)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Session token issued by `POST /v1/auth/session`. Attach as `Authorization: Bearer <token>`."
  schemas:
...
    QuotaExceeded:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            - quota_exceeded
        remaining:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
        priceUsd:
          type: number
        interval:
          type: string
          enum:
            - month
      required:
        - error
        - code
        - remaining
        - limit
...
    RivalriesResponse:
      type: object
      properties:
        rivalries:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              ticker:
                type: string
              rivalTicker:
                type: string
              wins:
                type: number
              losses:
                type: number
              draws:
                type: number
              currentPick:
                type: string
                enum:
                  - ticker
                  - rival
              weekStart:
                type: string
              createdAt:
                type: string
            required:
              - id
              - ticker
              - rivalTicker
              - wins
              - losses
              - draws
              - weekStart
              - createdAt
        count:
          type: number
      required:
        - rivalries
        - count
    CreateRivalryRequest:
      type: object
      properties:
        ticker:
          type: string
        rivalTicker:
          type: string
      required:
        - ticker
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
          required:
            - code
            - message
      required:
        - error
...
paths:
...
  /v1/rivalries:
    get:
      summary: Weekly matchups for the signed-in user
      description: "The caller's tracked solo matchups (e.g. NVDA vs AMD) with the running win/loss/draw record. Collection and comprehension mechanic only: a rivalry is not a position and the copy never instructs a trade. `weekStart` is the Monday of the current round as a UTC calendar day, so the round boundary does not move with the device timezone."
      tags:
        - identify
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Rivalries returned newest-first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RivalriesResponse"
        "400":
          description: Bad request — validation failed against the zod schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "429":
          description: Rate limit exceeded (per-user or per-ip).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Start a weekly matchup against a comparable
      description: Creates a solo weekly matchup for one of the caller's finds. Omit `rivalTicker` and the server picks the opponent from the existing comparables pipeline in `packages/finance` — the client must not invent a rival, and a ticker with no resolvable comparable returns 422 rather than a fabricated opponent. The optional pre-registered `currentPick` on the returned rivalry is a conviction game scored for XP; it is not an order and creates no position.
      tags:
        - identify
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRivalryRequest"
      responses:
        "201":
          description: Rivalry created; the full updated list is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RivalriesResponse"
        "400":
          description: Bad request — validation failed against the zod schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "409":
          description: A rivalry for this exact pairing already exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "422":
          description: No comparable could be resolved for `ticker` — pass an explicit `rivalTicker` instead.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "429":
          description: Rate limit exceeded (per-user or per-ip).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/rivalries/{id}/pick:
    post:
      summary: Pre-register (or clear) the conviction pick for the open round
      description: 'Body is `{ pick: "ticker" | "rival" | null }` — `null` clears the pick. A correct pick earns XP at the weekly close, which then clears it so a pick never carries into a round it was not registered for. Not an order; creates no position.'
      tags:
        - identify
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pick:
                  type:
                    - string
                    - "null"
                  enum:
                    - ticker
                    - rival
              required:
                - pick
      responses:
        "200":
          description: Pick recorded; the full updated list is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RivalriesResponse"
        "400":
          description: Bad request — validation failed against the zod schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "404":
          description: Rivalry not found (or not the caller's).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "429":
          description: Rate limit exceeded (per-user or per-ip).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/rivalries/{id}:
    delete:
      summary: Delete a rivalry
      tags:
        - identify
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "204":
          description: Rivalry deleted.
        "400":
          description: Bad request — validation failed against the zod schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "404":
          description: Rivalry not found (or not the caller's).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "429":
          description: Rate limit exceeded (per-user or per-ip).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...